refactored map.js, made prototype of pop-up window for bus markers
