+++ client/src/map.js
@@ -1,127 +1,138 @@
-import React from 'react';
-import GoogleMapReact from 'google-map-react';
-import Fab from '@material-ui/core/Fab';
-import Brightness3Icon from '@material-ui/icons/Brightness3';
-import MyLocationIcon from '@material-ui/icons/MyLocation';
-import PersonPinCircleIcon from '@material-ui/icons/PersonPinCircle';
-import './App.css';
+import React from "react";
+import { useState } from "react";
+import GoogleMapReact from "google-map-react";
+import Fab from "@material-ui/core/Fab";
+import Paper from "@material-ui/core/Paper";
+import Brightness3Icon from "@material-ui/icons/Brightness3";
+import MyLocationIcon from "@material-ui/icons/MyLocation";
+import PersonPinCircleIcon from "@material-ui/icons/PersonPinCircle";
+import "./App.css";
 
+function Map(props) {
 
-const defaultLat = 59.8585;
-const defaultLng = 17.6389;
-const styles = require('./mapstyle.json');
-const Marker = () => <div className='marker'>&#128652;</div>;
-const CenterMarker = () => <div className='centerMarker'><PersonPinCircleIcon color="secondary"/></div>;
+  const defaultLat = 59.8585;
+  const defaultLng = 17.6389;
+  const defaultCenter = {
+    lat: defaultLat,
+    lng: defaultLng,
+  };
+  const styles = require("./mapstyle.json");
+  const Marker = () => <div className="marker">&#128652;</div>;
+  const CenterMarker = () => (
+    <div className="centerMarker">
+      <PersonPinCircleIcon color="secondary" />
+    </div>
+  );
 
-class Map extends React.Component {
-
-	constructor(props) {
-		super(props);
-		this.state = {
-			realtimeData: props.realtimeData, // realtime information about the busses
-      theme: styles.day,
-      center: {
-        lat: defaultLat,
-        lng: defaultLng
-      }
-		};
-    this.updateLocation = this.updateLocation.bind(this);
-    this.setCoordinates = this.setCoordinates.bind(this);
-	};
-
-  changeTheme() {
-    if (this.state.theme === styles.day) {
-      this.setState({
-        realtimeData: this.state.realtimeData, // realtime information about the busses
-        theme: styles.night
-      });
-    } else {
-      this.setState({
-        realtimeData: this.state.realtimeData, // realtime information about the busses
-        theme: styles.day
-      });
-    }
+  const displayInfo = (obj) => {
+    setBusInfo(<Paper 
+      id="busInfo"
+      lat={obj.position.latitude} 
+      lng={obj.position.longitude}>
+        bus {obj.id}
+      </Paper>);
   };
 
-  updateLocation() {
+  const clearInfo = () => {
+    setBusInfo('');
+  };
+
+  const [busInfo, setBusInfo] = useState();
+  const [currentTheme, setCurrentTheme] = useState(styles.day);
+  const [realtimeData, setRealtimeData] = useState(props.realtimeData);
+  const [currentCenter, setCurrentCenter] = useState(defaultCenter);
+  
+  
+  const updateLocation = () => {
     if(navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition(this.setCoordinates) 
+      navigator.geolocation.getCurrentPosition(setCoordinates);
     } else {
-      alert("Browser error")
+      alert("Browser error");
     }
   };
 
-  setCoordinates(position) {
-    this.setState({
-      center: {
-        lat: position.coords.latitude,
-        lng: position.coords.longitude
+  const setCoordinates = (position) => {
+    setCurrentCenter({
+      lat: position.coords.latitude,
+      lng: position.coords.longitude
+    });
+  };
+
+  const changeTheme = () => {
+    if (currentTheme === styles.day) {
+        setCurrentTheme(styles.night)
+    } else {
+        setCurrentTheme(styles.day)
+    }
+  }
+
+  var arr = [
+    {
+      id: 1,
+      position: {
+        latitude:  59.8595,
+        longitude: 17.6389
       }
-    })
-  };
-
-
-	static defaultProps = {
-		// Coordinates of Stora torget, Uppsala
-    center: {
-      lat: defaultLat,
-      lng: defaultLng
     },
-    zoom: 14
-  };
+    {
+      id: 2,
+      position: {
+        latitude: 59.8575,
+        longitude: 17.6399
+      }
+    }
+  ];
 
-	componentDidMount() {};
-
-
-
-	render() {
-    
-		return (
-			<div style={{ height: '100vh', width: '100%' }}>
-        <GoogleMapReact
-          bootstrapURLKeys={{ key: '' }}
-          defaultCenter={this.props.center}
-          defaultZoom={this.props.zoom}
-          center={this.state.center}
-					options={{
-					 styles: this.state.theme,
-					 disableDefaultUI: true,
-					 fullscreenControl: false
-				 	}}
-        >
-          {
-            this.state.realtimeData.map((obj, i) => (
-             <Marker
-                key={obj.id}
-                lat={obj.vehicle.position.latitude}
-                lng={obj.vehicle.position.longitude}
-
-              />
-            ))
-          }
-
-          <CenterMarker
-            lat = {this.state.center.lat}
-            lng = {this.state.center.lng}
-          />
+  return (
+    <div style={{ height: "100vh", width: "100%" }}>
+      <GoogleMapReact
+        onClick={clearInfo}
+        bootstrapURLKeys={{ key: "" }}
+        defaultCenter={defaultCenter}
+        defaultZoom={15}
+        center={currentCenter}
+        options={{
+          styles: currentTheme,
+          disableDefaultUI: true,
+          fullscreenControl: false,
+        }}
+      >
+        {arr.map((obj, i) => (
+          <div 
+          className="marker"
+          key={obj.id}
+          lat={obj.position.latitude}
+          lng={obj.position.longitude}
+          onClick= {() => {displayInfo(obj)}}
+          >
+            &#128652;
+          </div>
           
-        </GoogleMapReact>
-        <Fab
-	        	id = "locationButton"
-	        	color = "primary"
-	        	aria-label = "locationButton"
-            onClick = {this.updateLocation}>
-	          <MyLocationIcon/>
-        </Fab>
-        <Fab 
-          color = "primary" 
-          id = "themeButton" 
-          onClick = {() => this.changeTheme()}>
-          <Brightness3Icon/>
-        </Fab>
-      </div>
-		);
-	};
+        ))}
+        {busInfo}
+        <CenterMarker 
+        lat={currentCenter.lat}
+        lng={currentCenter.lng}
+         />
+
+      </GoogleMapReact>
+      <Fab
+        id="locationButton"
+        color="primary"
+        aria-label="locationButton"
+        onClick={updateLocation}
+      >
+        <MyLocationIcon />
+      </Fab>
+      <Fab
+        color="primary"
+        id="themeButton"
+        onClick = {changeTheme}
+      >
+        <Brightness3Icon />
+      </Fab>
+    </div>
+  );
 }
 
 export default Map;
